fix: honor custom -Dpackage in generated projects
Passing -Dpackage=com.foo produced a source tree whose directories did not match
the package declarations: files landed under {package}/{artifactIdNoSpecialCharacters}/...
while declaring package ${groupId}.${artifactIdNoSpecialCharacters}.* — an extra path
segment (compiles, but IDEs and checkstyle flag every file).

Root cause: the sub-archetype Java lived under a literal __artifactIdNoSpecialCharacters__
directory inside a packaged=true fileSet (which already prepends ${package}), and the
package/import/classpath references were hardcoded to ${groupId}.${artifactIdNoSpecialCharacters}.

Fix:
- Flatten the __artifactIdNoSpecialCharacters__ path segment from the api/core/application
  Java source trees (the packaged fileSet supplies the package dirs).
- Replace ${groupId}.${artifactIdNoSpecialCharacters} with ${package} everywhere it named
  the project package: Java package/imports, the datasource classPath, the versioned-ui
  override HTL, and the pom rootPackage properties.
- In the meta-archetype groovy, default package to ${groupId}.${artifactIdNoSpecialCharacters}
  when the user did not pass -Dpackage, so the default layout is byte-identical while a
  custom -Dpackage is honored verbatim.
- Pin the conventional package in each sub-archetype IT's archetype.properties so their
  reference projects are unchanged.

Verified: default generation unchanged (composite IT green); -Dpackage=org.acme.custom
produces org/acme/custom/... with matching declarations, classPath and HTL, and builds green.
