--- kestros-api-archetype/src/main/resources/archetype-resources/src/test/java/__artifactIdNoSpecialCharacters__/api/exceptions/SampleModelRetrievalExceptionTest.java
@@ -1,31 +1,0 @@
-package ${groupId}.${artifactIdNoSpecialCharacters}.api.exceptions;
-
-import junit.framework.TestCase;
-import org.junit.Test;
-
-public class SampleModelRetrievalExceptionTest extends TestCase {
-
-  private SampleModelRetrievalException sampleModelRetrievalException;
-
-  @Test
-  public void testSampleModelRetrievalException() {
-    sampleModelRetrievalException = new SampleModelRetrievalException("message");
-    assertEquals("message", sampleModelRetrievalException.getMessage());
-  }
-
-  @Test
-  public void testSampleModelRetrievalExceptionWithCause() {
-    Throwable cause = new Throwable();
-    sampleModelRetrievalException = new SampleModelRetrievalException("message", cause);
-    assertEquals("message", sampleModelRetrievalException.getMessage());
-    assertEquals(cause, sampleModelRetrievalException.getCause());
-  }
-
-  @Test
-  public void testSampleModelRetrievalExceptionWithThrowable() {
-    Throwable cause = new Throwable();
-    sampleModelRetrievalException = new SampleModelRetrievalException(cause);
-    assertEquals(cause, sampleModelRetrievalException.getCause());
-  }
-
-}
