+++ kestros-api-archetype/src/test/resources/projects/it2/archetype.properties
@@ -4,7 +4,7 @@ version=0.0.1-SNAPSHOT
 artifactIdNoSpecialCharacters=buildintegrationtest2
 artifactDescription=Build Integration Test2
 organizationName=Test Organization2
-package=com.test2
+package=com.test2.buildintegrationtest2
 artifactName=Build Integration Test2
 hasParentProject=false
 hashSymbol=#
