+++ kestros-application-archetype/src/main/resources/archetype-resources/pom.xml
@@ -30,7 +30,7 @@
     <packaging>bundle</packaging>
 
     <properties>
-        <rootPackage>${groupId}.${artifactIdNoSpecialCharacters}.application</rootPackage>
+        <rootPackage>${package}.application</rootPackage>
 #if ($hasParentProject == "true")
         <bundleCategory>${organization.name}</bundleCategory>
 #else
