+++ kestros-application-archetype/src/main/resources/archetype-resources/src/content/jcr_root/apps/kestros/commons/components/lists/card-list/__artifactIdShorthand__-versioned-ui/versions/0.0.1/content.html
@@ -9,7 +9,7 @@
     the sample cards. Compare with the sample site page, which uses the UNVERSIONED framework and
     selects the datasource by property. Delete this file to fall back to the common view and
     author-selected datasources. */-->
-<sly data-sly-use.cardList="${sign}{'${groupId}.${artifactIdNoSpecialCharacters}.core.datasources.SampleCardListDataSource'}"/>
+<sly data-sly-use.cardList="${sign}{'${package}.core.datasources.SampleCardListDataSource'}"/>
 
 <div class="card-list card-list--sample" data-sly-list.card="${sign}{cardList.cards}">
     <sly data-sly-resource="${sign}{card}"/>
