+++ kestros-application-archetype/src/main/resources/archetype-resources/src/content/jcr_root/apps/kestros/commons/components/lists/card-list/datasources/__artifactId__-sample-cards/.content.xml
@@ -4,5 +4,5 @@
           jcr:title="${artifactName} Sample Cards"
           jcr:description="Cards built from the Sample Models provided by this project's SampleService OSGi service. Select this datasource on a Card List component (stored on the component instance as the kes:datasource property)."
           group="${artifactName}"
-          classPath="${groupId}.${artifactIdNoSpecialCharacters}.core.datasources.SampleCardListDataSource"
+          classPath="${package}.core.datasources.SampleCardListDataSource"
           container="{Boolean}true"/>
