+++ kestros-content-archetype/src/main/resources/archetype-resources/pom.xml
@@ -30,7 +30,7 @@
     <packaging>bundle</packaging>
 
     <properties>
-        <rootPackage>${groupId}.${artifactIdNoSpecialCharacters}</rootPackage>
+        <rootPackage>${package}</rootPackage>
 #if ($hasParentProject == "true")
         <bundleCategory>${organization.name}</bundleCategory>
 #else
