--- kestros-core-archetype/src/main/resources/archetype-resources/src/test/java/__artifactIdNoSpecialCharacters__/core/models/SampleModelImplTest.java
@@ -1,22 +1,0 @@
-package ${groupId}.${artifactIdNoSpecialCharacters}.core.models;
-
-import static org.junit.Assert.assertEquals;
-
-import org.junit.Test;
-
-public class SampleModelImplTest {
-
-  private SampleModelImpl sampleModelImpl;
-
-  @Test
-  public void testGetTitle() {
-    sampleModelImpl = new SampleModelImpl("title", "description");
-    assertEquals("title", sampleModelImpl.getTitle());
-  }
-
-  @Test
-  public void testGetDescription() {
-    sampleModelImpl = new SampleModelImpl("title", "description");
-    assertEquals("description", sampleModelImpl.getDescription());
-  }
-}
