+++ kestros-core-archetype/src/test/resources/projects/it3/archetype.properties
@@ -4,7 +4,7 @@ version=0.0.1-SNAPSHOT
 artifactIdNoSpecialCharacters=buildintegrationtest
 artifactDescription=Build Integration Test
 organizationName=Test Organization
-package=com.test
+package=com.test.buildintegrationtest
 artifactName=Build Integration Test
 hasParentProject=true
 hashSymbol=#
